Basic nearby logic implemented
Computer will select nearby squares when hit. Implements
a stack method to try and find the ship if it gets off target.

However, the logic isn't the smartest as it will not account for
direction.

Also, if a ship IS sunk, the entire logic goes back to random.
This means if a ship is sunk while hitting another ship, the
previous hit memory is lost.

Some styling changes also added in terms of hovering over a square.

Last changes will be to style it up a bit!